move tracetmc tests to their own file
